refactor(keychain): move keychain polyfill to separate file
