+++ tsconfig.json
@@ -43,7 +43,7 @@
     "baseUrl": "./" /* Base directory to resolve non-absolute module names. */,
     "paths": {
       /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
-      "src/lib/keychain": ["./src/node.polyfill"],
+      "src/lib/keychain": ["./src/polyfills/keychain"],
       "rn-fetch-blob": ["./src/node.polyfill"],
       "src/lib/errors/sentry": ["./src/node.polyfill"],
       "react-native-localize": ["./src/node.polyfill"],
